+++ src/compiled_template.php
@@ -1,0 +1,14 @@
+class <?php echo $compileClass; ?> extends <?php echo $compileParentClass; ?> 
+{
+    const METHOD_MAPPING = <?php var_export($this->entryToMethods); ?>;
+    
+<?php foreach ($this->methods as $methodName => $content): ?>
+    protected function <?php echo $methodName; ?>()
+    {
+        <?php echo $content; ?>
+        
+    }
+    
+<?php endforeach; ?>
+
+}